# internal/types/types.go
// Package types provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package types

import (
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// AuthResponse defines model for AuthResponse.
type AuthResponse struct {
	// AccessToken JWT access token
	AccessToken *string `json:"access_token,omitempty"`

	// ExpiresIn Token expiration time in seconds
	ExpiresIn *int `json:"expires_in,omitempty"`

	// RefreshToken Refresh token
	RefreshToken *string `json:"refresh_token,omitempty"`

	// TokenType Token type
	TokenType *string       `json:"token_type,omitempty"`
	User      *UserResponse `json:"user,omitempty"`
}

// ChangePasswordRequest defines model for ChangePasswordRequest.
type ChangePasswordRequest struct {
	// CurrentPassword Current password
	CurrentPassword string `json:"current_password"`

	// NewPassword New password
	NewPassword string `json:"new_password"`
}

// CreateUserRequest defines model for CreateUserRequest.
type CreateUserRequest struct {
	// Email User's email address
	Email openapi_types.Email `json:"email"`

	// Name User's full name
	Name string `json:"name"`

	// Password User's password
	Password string `json:"password"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Code HTTP status code
	Code *int `json:"code,omitempty"`

	// Error Error type
	Error *string `json:"error,omitempty"`

	// Message Error message
	Message *string `json:"message,omitempty"`

	// Success Whether the operation was successful
	Success *bool `json:"success,omitempty"`

	// Timestamp Response timestamp
	Timestamp *time.Time `json:"timestamp,omitempty"`
}

// ForgotPasswordRequest defines model for ForgotPasswordRequest.
type ForgotPasswordRequest struct {
	// Email User's email address
	Email openapi_types.Email `json:"email"`
}

// HealthData defines model for HealthData.
type HealthData struct {
	// Services Health status of individual services
	Services *map[string]string `json:"services,omitempty"`

	// Status Overall health status
	Status *string `json:"status,omitempty"`

	// Version Application version
	Version *string `json:"version,omitempty"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	// Email User's email address
	Email openapi_types.Email `json:"email"`

	// Password User's password
	Password string `json:"password"`
}

// PaginatedUserResponse defines model for PaginatedUserResponse.
type PaginatedUserResponse struct {
	// Code HTTP status code
	Code *int              `json:"code,omitempty"`
	Data *UserListResponse `json:"data,omitempty"`

	// Message Response message
	Message    *string         `json:"message,omitempty"`
	Pagination *PaginationInfo `json:"pagination,omitempty"`

	// Success Whether the operation was successful
	Success *bool `json:"success,omitempty"`

	// Timestamp Response timestamp
	Timestamp *time.Time `json:"timestamp,omitempty"`
}

// PaginationInfo defines model for PaginationInfo.
type PaginationInfo struct {
	// HasNext Whether there is a next page
	HasNext *bool `json:"has_next,omitempty"`

	// HasPrev Whether there is a previous page
	HasPrev *bool `json:"has_prev,omitempty"`

	// Limit Number of items per page
	Limit *int `json:"limit,omitempty"`

	// Page Current page number
	Page *int `json:"page,omitempty"`

	// Total Total number of items
	Total *int `json:"total,omitempty"`

	// TotalPages Total number of pages
	TotalPages *int `json:"total_pages,omitempty"`
}

// RefreshRequest defines model for RefreshRequest.
type RefreshRequest struct {
	// RefreshToken Refresh token
	RefreshToken string `json:"refresh_token"`
}

// RegisterRequest defines model for RegisterRequest.
type RegisterRequest struct {
	// Email User's email address
	Email openapi_types.Email `json:"email"`

	// Name User's full name
	Name string `json:"name"`

	// Password User's password
	Password string `json:"password"`
}

// ResetPasswordRequest defines model for ResetPasswordRequest.
type ResetPasswordRequest struct {
	// Email User's email address
	Email openapi_types.Email `json:"email"`

	// NewPassword New password
	NewPassword string `json:"new_password"`

	// Token Password reset token
	Token string `json:"token"`
}

// StandardResponse defines model for StandardResponse.
type StandardResponse struct {
	// Code HTTP status code
	Code *int `json:"code,omitempty"`

	// Data Response data (varies by endpoint)
	Data *StandardResponse_Data `json:"data,omitempty"`

	// Error Error type (only present on errors)
	Error *string `json:"error,omitempty"`

	// Message Response message
	Message *string `json:"message,omitempty"`

	// Success Whether the operation was successful
	Success *bool `json:"success,omitempty"`

	// Timestamp Response timestamp
	Timestamp *time.Time `json:"timestamp,omitempty"`
}

// StandardResponse_Data Response data (varies by endpoint)
type StandardResponse_Data struct {
	union json.RawMessage
}

// UpdateUserRequest defines model for UpdateUserRequest.
type UpdateUserRequest struct {
	// Email User's email address
	Email openapi_types.Email `json:"email"`

	// Name User's full name
	Name string `json:"name"`

	// Password User's password
	Password string `json:"password"`
}

// UserListResponse defines model for UserListResponse.
type UserListResponse struct {
	// Users List of users
	Users *[]UserResponse `json:"users,omitempty"`
}

// UserResponse defines model for UserResponse.
type UserResponse struct {
	// CreatedAt User creation timestamp
	CreatedAt *time.Time `json:"created_at,omitempty"`

	// Email User's email address
	Email *openapi_types.Email `json:"email,omitempty"`

	// Id User ID
	Id *openapi_types.UUID `json:"id,omitempty"`

	// IsActive Whether the user account is active
	IsActive *bool `json:"is_active,omitempty"`

	// Name User's full name
	Name *string `json:"name,omitempty"`

	// UpdatedAt User last update timestamp
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
}

// GetApiV1UsersParams defines parameters for GetApiV1Users.
type GetApiV1UsersParams struct {
	// Limit Number of users to return (default: 10, max: 100)
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Number of users to skip (default: 0)
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// PutApiV1AuthChangePasswordJSONRequestBody defines body for PutApiV1AuthChangePassword for application/json ContentType.
type PutApiV1AuthChangePasswordJSONRequestBody = ChangePasswordRequest

// PostApiV1AuthForgotPasswordJSONRequestBody defines body for PostApiV1AuthForgotPassword for application/json ContentType.
type PostApiV1AuthForgotPasswordJSONRequestBody = ForgotPasswordRequest

// PostApiV1AuthLoginJSONRequestBody defines body for PostApiV1AuthLogin for application/json ContentType.
type PostApiV1AuthLoginJSONRequestBody = LoginRequest

// PostApiV1AuthRefreshJSONRequestBody defines body for PostApiV1AuthRefresh for application/json ContentType.
type PostApiV1AuthRefreshJSONRequestBody = RefreshRequest

// PostApiV1AuthRegisterJSONRequestBody defines body for PostApiV1AuthRegister for application/json ContentType.
type PostApiV1AuthRegisterJSONRequestBody = RegisterRequest

// PostApiV1AuthResetPasswordJSONRequestBody defines body for PostApiV1AuthResetPassword for application/json ContentType.
type PostApiV1AuthResetPasswordJSONRequestBody = ResetPasswordRequest

// PostApiV1UsersJSONRequestBody defines body for PostApiV1Users for application/json ContentType.
type PostApiV1UsersJSONRequestBody = CreateUserRequest

// PutApiV1UsersIdJSONRequestBody defines body for PutApiV1UsersId for application/json ContentType.
type PutApiV1UsersIdJSONRequestBody = UpdateUserRequest

// AsUserResponse returns the union data inside the StandardResponse_Data as a UserResponse
func (t StandardResponse_Data) AsUserResponse() (UserResponse, error) {
	var body UserResponse
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromUserResponse overwrites any union data inside the StandardResponse_Data as the provided UserResponse
func (t *StandardResponse_Data) FromUserResponse(v UserResponse) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeUserResponse performs a merge with any union data inside the StandardResponse_Data, using the provided UserResponse
func (t *StandardResponse_Data) MergeUserResponse(v UserResponse) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

// AsUserListResponse returns the union data inside the StandardResponse_Data as a UserListResponse
func (t StandardResponse_Data) AsUserListResponse() (UserListResponse, error) {
	var body UserListResponse
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromUserListResponse overwrites any union data inside the StandardResponse_Data as the provided UserListResponse
func (t *StandardResponse_Data) FromUserListResponse(v UserListResponse) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeUserListResponse performs a merge with any union data inside the StandardResponse_Data, using the provided UserListResponse
func (t *StandardResponse_Data) MergeUserListResponse(v UserListResponse) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

// AsPaginatedUserResponse returns the union data inside the StandardResponse_Data as a PaginatedUserResponse
func (t StandardResponse_Data) AsPaginatedUserResponse() (PaginatedUserResponse, error) {
	var body PaginatedUserResponse
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromPaginatedUserResponse overwrites any union data inside the StandardResponse_Data as the provided PaginatedUserResponse
func (t *StandardResponse_Data) FromPaginatedUserResponse(v PaginatedUserResponse) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergePaginatedUserResponse performs a merge with any union data inside the StandardResponse_Data, using the provided PaginatedUserResponse
func (t *StandardResponse_Data) MergePaginatedUserResponse(v PaginatedUserResponse) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

// AsHealthData returns the union data inside the StandardResponse_Data as a HealthData
func (t StandardResponse_Data) AsHealthData() (HealthData, error) {
	var body HealthData
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromHealthData overwrites any union data inside the StandardResponse_Data as the provided HealthData
func (t *StandardResponse_Data) FromHealthData(v HealthData) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeHealthData performs a merge with any union data inside the StandardResponse_Data, using the provided HealthData
func (t *StandardResponse_Data) MergeHealthData(v HealthData) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

// AsAuthResponse returns the union data inside the StandardResponse_Data as a AuthResponse
func (t StandardResponse_Data) AsAuthResponse() (AuthResponse, error) {
	var body AuthResponse
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromAuthResponse overwrites any union data inside the StandardResponse_Data as the provided AuthResponse
func (t *StandardResponse_Data) FromAuthResponse(v AuthResponse) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeAuthResponse performs a merge with any union data inside the StandardResponse_Data, using the provided AuthResponse
func (t *StandardResponse_Data) MergeAuthResponse(v AuthResponse) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

func (t StandardResponse_Data) MarshalJSON() ([]byte, error) {
	b, err := t.union.MarshalJSON()
	return b, err
}

func (t *StandardResponse_Data) UnmarshalJSON(b []byte) error {
	err := t.union.UnmarshalJSON(b)
	return err
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Change password
	// (PUT /api/v1/auth/change-password)
	PutApiV1AuthChangePassword(c *gin.Context)
	// Request password reset
	// (POST /api/v1/auth/forgot-password)
	PostApiV1AuthForgotPassword(c *gin.Context)
	// Login user
	// (POST /api/v1/auth/login)
	PostApiV1AuthLogin(c *gin.Context)
	// Logout user
	// (POST /api/v1/auth/logout)
	PostApiV1AuthLogout(c *gin.Context)
	// Refresh access token
	// (POST /api/v1/auth/refresh)
	PostApiV1AuthRefresh(c *gin.Context)
	// Register a new user
	// (POST /api/v1/auth/register)
	PostApiV1AuthRegister(c *gin.Context)
	// Reset password
	// (POST /api/v1/auth/reset-password)
	PostApiV1AuthResetPassword(c *gin.Context)
	// Health check
	// (GET /api/v1/health)
	GetApiV1Health(c *gin.Context)
	// Get metrics
	// (GET /api/v1/metrics)
	GetApiV1Metrics(c *gin.Context)
	// List users
	// (GET /api/v1/users)
	GetApiV1Users(c *gin.Context, params GetApiV1UsersParams)
	// Create user
	// (POST /api/v1/users)
	PostApiV1Users(c *gin.Context)
	// Get current user profile
	// (GET /api/v1/users/me)
	GetApiV1UsersMe(c *gin.Context)
	// Delete user
	// (DELETE /api/v1/users/{id})
	DeleteApiV1UsersId(c *gin.Context, id openapi_types.UUID)
	// Get user by ID
	// (GET /api/v1/users/{id})
	GetApiV1UsersId(c *gin.Context, id openapi_types.UUID)
	// Update user
	// (PUT /api/v1/users/{id})
	PutApiV1UsersId(c *gin.Context, id openapi_types.UUID)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// PutApiV1AuthChangePassword operation middleware
func (siw *ServerInterfaceWrapper) PutApiV1AuthChangePassword(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PutApiV1AuthChangePassword(c)
}

// PostApiV1AuthForgotPassword operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1AuthForgotPassword(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostApiV1AuthForgotPassword(c)
}

// PostApiV1AuthLogin operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1AuthLogin(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostApiV1AuthLogin(c)
}

// PostApiV1AuthLogout operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1AuthLogout(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostApiV1AuthLogout(c)
}

// PostApiV1AuthRefresh operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1AuthRefresh(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostApiV1AuthRefresh(c)
}

// PostApiV1AuthRegister operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1AuthRegister(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostApiV1AuthRegister(c)
}

// PostApiV1AuthResetPassword operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1AuthResetPassword(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostApiV1AuthResetPassword(c)
}

// GetApiV1Health operation middleware
func (siw *ServerInterfaceWrapper) GetApiV1Health(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetApiV1Health(c)
}

// GetApiV1Metrics operation middleware
func (siw *ServerInterfaceWrapper) GetApiV1Metrics(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetApiV1Metrics(c)
}

// GetApiV1Users operation middleware
func (siw *ServerInterfaceWrapper) GetApiV1Users(c *gin.Context) {

	var err error

	c.Set(BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetApiV1UsersParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", c.Request.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", c.Request.URL.Query(), &params.Offset)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter offset: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetApiV1Users(c, params)
}

// PostApiV1Users operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1Users(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostApiV1Users(c)
}

// GetApiV1UsersMe operation middleware
func (siw *ServerInterfaceWrapper) GetApiV1UsersMe(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetApiV1UsersMe(c)
}

// DeleteApiV1UsersId operation middleware
func (siw *ServerInterfaceWrapper) DeleteApiV1UsersId(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteApiV1UsersId(c, id)
}

// GetApiV1UsersId operation middleware
func (siw *ServerInterfaceWrapper) GetApiV1UsersId(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetApiV1UsersId(c, id)
}

// PutApiV1UsersId operation middleware
func (siw *ServerInterfaceWrapper) PutApiV1UsersId(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PutApiV1UsersId(c, id)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.PUT(options.BaseURL+"/api/v1/auth/change-password", wrapper.PutApiV1AuthChangePassword)
	router.POST(options.BaseURL+"/api/v1/auth/forgot-password", wrapper.PostApiV1AuthForgotPassword)
	router.POST(options.BaseURL+"/api/v1/auth/login", wrapper.PostApiV1AuthLogin)
	router.POST(options.BaseURL+"/api/v1/auth/logout", wrapper.PostApiV1AuthLogout)
	router.POST(options.BaseURL+"/api/v1/auth/refresh", wrapper.PostApiV1AuthRefresh)
	router.POST(options.BaseURL+"/api/v1/auth/register", wrapper.PostApiV1AuthRegister)
	router.POST(options.BaseURL+"/api/v1/auth/reset-password", wrapper.PostApiV1AuthResetPassword)
	router.GET(options.BaseURL+"/api/v1/health", wrapper.GetApiV1Health)
	router.GET(options.BaseURL+"/api/v1/metrics", wrapper.GetApiV1Metrics)
	router.GET(options.BaseURL+"/api/v1/users", wrapper.GetApiV1Users)
	router.POST(options.BaseURL+"/api/v1/users", wrapper.PostApiV1Users)
	router.GET(options.BaseURL+"/api/v1/users/me", wrapper.GetApiV1UsersMe)
	router.DELETE(options.BaseURL+"/api/v1/users/:id", wrapper.DeleteApiV1UsersId)
	router.GET(options.BaseURL+"/api/v1/users/:id", wrapper.GetApiV1UsersId)
	router.PUT(options.BaseURL+"/api/v1/users/:id", wrapper.PutApiV1UsersId)
}